test: storybook file for CategoryButtonContainer
